requirements.txt
flask
mlflow
notebook
pandas
numpy
matplotlib
python-box
tensorflow
scikit-learn
tqdm
pyYAML
ensure
joblib
types-pyYAML
Flask-Cors
dvc
dvc-s3
dagshub